[USB Device] The USB Device Layer data structure has been updated to support multi transaction control OUT transfers. [MH3-81446]

+++ middleware/src/usb_device_local.h
@@ -114,8 +114,8 @@ typedef struct
 {
     /* Control transfer state */
     bool  inProgress;
-    uint8_t rxDataCount;
-    uint8_t expectedRxDataCount;
+    uint32_t rxDataCount;
+    uint32_t expectedRxDataCount;
     SYS_MODULE_INDEX handlerIndex;
     uint8_t * rxBuffer;
